Only allow 1 active schedule within date range, fix auth for schedule…

+++ server/src/main/java/com/projects/shiftproscheduler/applicationuser/ApplicationUserRepository.java
@@ -4,6 +4,6 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Long> {
+public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Integer> {
     Optional<ApplicationUser> findByUsername(String username);
 }
